financeapi/requirements.txt
django 
django-rest-framework
mysqlclient==1.3.13
django-mysql==2.2.0
